refactor: reduce number of files
